# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estudy",
        "description": "API doc for Estudy",
        "contact": {
            "url": "https://justice-chimobi.vercel.app/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
...
        "/courses/m/all": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get all courses created by a mentor",
                "description": "Get all courses created by a mentor",
                "operationId": "e5692be9b171028f656d54a5174377c1",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "estudy"
                                                },
                                                "price": {
                                                    "type": "string",
                                                    "format": "currency",
                                                    "example": "20.00"
                                                },
                                                "video": {
                                                    "type": "numner",
                                                    "example": "2"
                                                },
                                                "created_at": {
                                                    "$ref": "#/components/schemas/DateTimeResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/courses/m/enrolled-students": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get all students enrolled on a mentor course",
                "description": "Get all students enrolled on a mentor course",
                "operationId": "98a26bcd4829f4d863102fe9e3db4e38",
                "responses": {
                    "200": {
                        "description": "course overview",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MentorEnrolledStudentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/courses/m/video": {
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Upload course video to cloudinary",
                "description": "Upload course video to cloudinary and get actual URL from response to use for course endpoint",
                "operationId": "PostMentorcourseVideo",
                "requestBody": {
                    "description": "Video file to upload max 8 mb",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "video": {
                                        "description": "Video file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "uploaded",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "uploaded",
                                    "video": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.mp4"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/courses/s/all": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get all courses for students dashboard",
                "description": "Get all courses for students dashboard",
                "operationId": "55d8b31020ceda2c0c79d86782b96e56",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "intro course"
                                                    },
                                                    "thumbnail": {
                                                        "type": "string",
                                                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "format": "currency",
                                                        "example": "20.00"
                                                    },
                                                    "profile": {
                                                        "$ref": "#/components/schemas/CourseProfileResource"
                                                    },
                                                    "isEnrolled": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "courseEnroll": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/CourseEnrollResource"
                                                        }
                                                    },
                                                    "created_at": {
                                                        "$ref": "#/components/schemas/DateTimeResource"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/courses/s/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get individual course on students dashboard",
                "description": "Get individual course on students dashboard",
                "operationId": "ed1373af8f197e957a7baf22b677f32a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "course overview",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentCourseOverviewResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/user/delete": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete user from database",
                "description": "Delete user from database",
                "operationId": "6c125472a3ed59db6b89bf70aae5610f",
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CourseEnrollResource": {
                "title": "CourseEnrollResource",
                "description": "Course students enrolled in ",
                "properties": {
                    "course_id": {
                        "description": "course identifier",
                        "type": "string",
                        "example": "230e8400-e29b-41d4-a716-446655440000"
                    },
                    "user_id": {
                        "description": "user identifier",
                        "type": "string",
                        "example": "660e8400-e29b-41d4-a716-156655440000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CourseEnrollResource"
                }
            },
            "CourseProfileOverviewResource": {
                "title": "CourseProfileOverviewResource",
                "description": "Course overview",
                "properties": {
                    "fullname": {
                        "description": "full name",
                        "type": "string",
                        "example": "Gift Owens"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CourseProfileOverviewResource"
                }
            },
            "CourseProfileResource": {
                "title": "CourseProfileResource",
                "description": "Course Profile Resource",
                "properties": {
                    "fullname": {
                        "description": "Full name",
                        "type": "string",
                        "example": "Gift Owens"
                    },
                    "avatar": {
                        "description": "Course name",
                        "type": "string",
                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CourseProfileResource"
                }
            },
            "DateTimeResource": {
                "title": "DateTimeResource",
                "description": "format date time",
                "properties": {
                    "human": {
                        "description": "long format string",
                        "type": "string",
                        "example": "2 hours ago"
                    },
                    "date_time": {
                        "description": "string representation",
                        "type": "string",
                        "example": "2024-01-22 12:01:58"
                    },
                    "human_short": {
                        "description": "short format string",
                        "type": "string",
                        "example": "2h ago"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DateTimeResource"
                }
            },
            "MentorEnrolledStudentResource": {
                "title": "MentorEnrolledStudentResource",
                "description": "get all students enrolled on mentor courses",
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "course_id": {
                        "description": "course id",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "owner_id": {
                        "description": "owner id",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "course_details": {
                        "description": "get course details name",
                        "type": "object",
                        "example": {
                            "course_name": "course name"
                        }
                    },
                    "enrolled_user_details": {
                        "$ref": "#/components/schemas/MentorStudentDetailsResource"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/DateTimeResource"
                    },
                    "pagination": {
                        "description": "get the paginated details",
                        "type": "object",
                        "example": {
                            "current_page": 1,
                            "last_page": 2,
                            "total": 8
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MentorEnrolledStudentResource"
                }
            },
            "MentorStudentDetailsResource": {
                "title": "MentorStudentDetailsResource",
                "description": "get students enrolled on mentor course details",
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "description": "student enroll name",
                        "type": "string",
                        "example": "Owens Joe"
                    },
                    "avatar": {
                        "description": "student enroll avatar",
                        "type": "string",
                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MentorStudentDetailsResource"
                }
            },
...
            "StudentCourseOverviewResource": {
                "title": "StudentCourseOverviewResource",
                "description": "course overview resource",
                "properties": {
                    "id": {
                        "description": "Course ID",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "description": "Course name",
                        "type": "string",
                        "example": "Intro Course"
                    },
                    "description": {
                        "description": "Course description",
                        "type": "string",
                        "example": "example of course description"
                    },
                    "title": {
                        "description": "Course title",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Introduction"
                        }
                    },
                    "video": {
                        "description": "Number of videos in the course",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "https://res.cloudinary.com/driggenlearn/video/upload/v1706030614/ayroiowquwwrgdogsznn.mp4"
                        }
                    },
                    "profile": {
                        "$ref": "#/components/schemas/CourseProfileOverviewResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StudentCourseOverviewResource"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}